requirements.txt
streamlit>=1.18.0
gTTS>=2.3.0
SpeechRecognition
Pillow
requests
